requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.11.0
pydantic==2.9.2
python-dotenv==1.0.1
jinja2==3.1.4
python-multipart==0.0.12
httpx==0.27.2
aiohttp==3.10.5
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0